pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "syspulse"
version = "0.1.0"
description = "Cross-platform security assessment agent"
requires-python = ">=3.8"
dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "eval_type_backport>=0.2.0",
    "typing_extensions>=4.0.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "jinja2>=3.1.0",
    "structlog>=24.1.0",
    "pywin32>=306; sys_platform=='win32'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pyinstaller>=6.4.0",
]

[project.scripts]
syspulse = "syspulse.cli:app"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
